refactor: move sq transformer to layout dir
